requirements.txt
pydantic~=1.8.2
uvicorn~=0.15.0
fastapi~=0.68.2
SQLAlchemy~=1.4.25
passlib~=1.7.4
python-dotenv~=0.19.0
python-jose~=3.3.0
python-multipart~=0.0.5